requirements.txt
click==7.1.2
python-jenkins==1.7.0
